Move changes from Layout to Banner
